+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- elaborator implicit-lambda insertion
 - fix elaborator: when build projection result type, make an error that Kernel reject it
 
 ## 0.5.1